+++ code/src/cljs/sixsq/nuvla/ui/utils/semantic_ui.cljs
@@ -187,6 +187,7 @@
 (def TextArea (r/adapt-react-class semantic/TextArea))
 
 (def Transition (r/adapt-react-class semantic/Transition))
+(def TransitionGroup (r/adapt-react-class semantic/TransitionGroup))
 
 (def TransitionablePortal (r/adapt-react-class semantic/TransitionablePortal))